chore(login): refactor the types package for organization and login
