+++ frontend/src/api/user/crud/delete.ts
@@ -2,7 +2,7 @@ import axios from 'api';
 import { ErrorResponseHandler } from 'api/ErrorResponseHandler';
 import { AxiosError } from 'axios';
 import { ErrorResponse, SuccessResponse } from 'types/api';
-import { PayloadProps, Props } from 'types/api/user/crud/deleteUser';
+import { PayloadProps, Props } from 'types/api/user/crud/delete';
 
 const deleteUser = async (
 	props: Props,
