+++ frontend/src/api/user/crud/get.ts
@@ -2,7 +2,7 @@ import axios from 'api';
 import { ErrorResponseHandlerV2 } from 'api/ErrorResponseHandlerV2';
 import { AxiosError } from 'axios';
 import { ErrorV2Resp, SuccessResponseV2 } from 'types/api';
-import { PayloadProps, UserResponse } from 'types/api/user/crud/getUser';
+import { PayloadProps, UserResponse } from 'types/api/user/crud/get';
 
 const getUser = async (): Promise<SuccessResponseV2<UserResponse>> => {
 	try {
