+++ frontend/src/api/user/crud/update.ts
@@ -2,7 +2,7 @@ import axios from 'api';
 import { ErrorResponseHandler } from 'api/ErrorResponseHandler';
 import { AxiosError } from 'axios';
 import { ErrorResponse, SuccessResponse } from 'types/api';
-import { PayloadProps, Props } from 'types/api/user/crud/editUser';
+import { PayloadProps, Props } from 'types/api/user/crud/update';
 
 const update = async (
 	props: Props,
