+++ frontend/src/api/user/preference/getPreference.ts
@@ -2,7 +2,7 @@ import axios from 'api';
 import { ErrorResponseHandler } from 'api/ErrorResponseHandler';
 import { AxiosError } from 'axios';
 import { ErrorResponse, SuccessResponse } from 'types/api';
-import { PayloadProps } from 'types/api/user/preference/getUserPreference';
+import { PayloadProps } from 'types/api/user/preference/get';
 
 const getPreference = async (): Promise<
 	SuccessResponse<PayloadProps> | ErrorResponse
