+++ frontend/src/container/OrganizationSettings/InviteUserModal/InviteUserModal.tsx
@@ -14,7 +14,7 @@ import {
 } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useQuery } from 'react-query';
-import { PayloadProps } from 'types/api/user/invite/getPendingInvites';
+import { PayloadProps } from 'types/api/user/invite/getAll';
 import { ROLES } from 'types/roles';
 
 import InviteTeamMembers from '../InviteTeamMembers';
