+++ frontend/src/pages/SignUp/index.tsx
@@ -1,5 +1,5 @@
 import { Typography } from 'antd';
-import getUserVersion from 'api/user/version/getVersion';
+import getUserVersion from 'api/user/version/get';
 import Spinner from 'components/Spinner';
 import { useAppContext } from 'providers/App/App';
 import { useTranslation } from 'react-i18next';
