--- frontend/src/types/api/organizations/getOrganization.ts
@@ -1,7 +1,0 @@
-export interface Organization {
-	createdAt: number;
-	id: string;
-	displayName: string;
-}
-
-export type PayloadProps = Organization[];
