+++ frontend/src/types/api/user/crud/delete.ts
