--- frontend/src/types/api/user/crud/getOrgMembers.ts
@@ -1,18 +1,0 @@
-import { ROLES } from 'types/roles';
-
-import { Organization } from '../../organizations/getOrganization';
-
-export interface Props {
-	orgId: Organization['id'];
-}
-
-interface OrgMembers {
-	createdAt: number;
-	email: string;
-	name: string;
-	role: ROLES;
-	token: string;
-	id: string;
-}
-
-export type PayloadProps = OrgMembers[];
