--- frontend/src/types/api/user/crud/updateRole.ts
@@ -1,10 +1,0 @@
-import { ROLES } from 'types/roles';
-
-export interface Props {
-	group_name: ROLES;
-	userId: string;
-}
-
-export interface PayloadProps {
-	data: string;
-}
