+++ frontend/src/types/api/user/factor_password/update.ts
