+++ frontend/src/types/api/user/invite/accept.ts
