+++ frontend/src/types/api/user/invite/delete.ts
