--- frontend/src/types/api/user/preference/setUserPreference.ts
@@ -1,8 +1,0 @@
-export interface Props {
-	isAnonymous: boolean;
-	hasOptedUpdates: boolean;
-}
-
-export interface PayloadProps {
-	data: string;
-}
